+++ s/core/tuples.ts
@@ -1,0 +1,13 @@
+
+
+export type Tuple2 = [x: number, y: number]
+export type Tuple3 = [x: number, y: number, z: number]
+export type Tuple4 = [x: number, y: number, z: number, w: number]
+
+export type Tuple16 = [
+	number, number, number, number,
+	number, number, number, number,
+	number, number, number, number,
+	number, number, number, number,
+]
+